Trim generated .travis.yml by fetching install script dynmically.
Not sure of the wisdom of this - one one hand we can fix things in planemo as time goes on and that is good and may prove needed since we are targeting galaxy-central. On the other hand this is less tweakable and may in fact be more likely to break - hard to say.

Example of workin generated .travis.yml file here https://github.com/jmchilton/bwa-mem/commit/c7a5133f1d3574079c40f8e2c20b9202fb8f72af.