Implement --update_test_data flag for test command.
When planemo test is called with this flag - the outputs of the tests are copied back into the files they are compared against in test-data. This is useful for updating these files automatically with small differences during minor updates.

Please review these changes carefully though and only use this option when you are sure the tool is correct - otherwise these tests are not useful for verifying correctness (only for reproducibility across systems) since they will just be comparing the tool against data the tool generated.