Friendly breakdown of tests (disable with --skip_summary).
The breakdown summary runs after testing and includes a list of each test and whether it passed or not (in festive green or red).

This feature of requires a feature (xunit output) in the latest development branch of Galaxy and so will not be compatible with the December 2014 release of Galaxy. Planemo will explicitly check Galaxy for xunit support and simply print a message about whether the test passed or failed if not available along with the path of the HTML output report which will contain the detailed breakdown. Even this degraded functionality is an improvement on planemo's test output however.

The commit also introduces a --test_output_xunit option - this in conjunction with --skip_summary is meant to be used with downstream testing infrastructure such as Jenkins - which can readily track testing statistics in xUnit format over time.