More intelligent test output linting.
Check for different kinds of expectations.

Testing to support this in the form of a directory of example tools - those that start with ``ok_`` will be linted during test with expectation that they pass and those starting with ``fail_`` will linted during testing with expectation that they fail.

Fixes #138 (the original problem anyway).