Allow the shed_ operations to consume git URLs instead of paths.
create, upload, download, diff, and lint projects directly from Github - closes #169.

Inputs starting with ``git+`` (e.g. git+/path/to/repo.git) with will have the ``git+`` stripped and passed directly to the ``git`` command-line. Alternatively URLs starting with `git:` will be passed as is to the ``git`` command-line. This seems to roughly correspond to ``pip``.