Look at file name only for implicit ignores.
Solved problem I had on a branch with './tool_dependencies.xml'
being ignored for a tool package because it allowed the exact
match 'tool_dependencies.xml' (likewise for tool suites).

On the other hand, with this change these special files don't
have to be at the top level of the Tool Shed tar-ball anymore.
Is a sub-folder acceptable?