Implement better tool file guessing.
Before it read the first 10 lines and look for <tool. This is bad for two reasons - one is it breaks if there are a several lines of comments at the beginning of the file and the second reason is that the point of limiting the number of lines is avoid loading huge files into memory - but XML files could easily have huge first lines (for instance results of unformatted web service calls).

This variant of guessing reads the first 5Kb in instead and still searches for <tool.

Fixes #231.