Implement profiles in sheds section of the ~/.planemo.yml.
Each shed section can now have a username which will be used in lieu of the top-level shed_username in that config. Each shed section can now specify a URL - existing profiles (toolshed, testtoolshed, and local) will be assigned URLs automatically if they are specified.

But new profiles can be created with arbitrary names and used in conjunction with shed_target.