dependency_script: cp rather than symlink to $DOWNLOAD_CACHE
When the downloaded file is an archive which gets unpacked in
the working directory, this is harmless. However, when the
download file is moved directly to $INSTALL_DIR (e.g. JAR files)
this leaves the installation with a symlink pointing at the
cache - which might get deleted.

Also, some tools will look for their own dependencies relative
to the real file (and not the symlink) which will point them
at $DOWNLOAD_CACHE rather than $INSTALL_DIR.