Implement ``clone`` and ``pull_request`` commands to ease PRs.
My vision of https://github.com/galaxyproject/planemo/issues/477.

```
$ planemo clone --branch bwa-fix tools-iuc
$ cd tools-iuc
$ # Make changes.
$ git add -p # Add desired changes.
$ git commit -m "Fix bwa problem."
$ planemo pull_request -m "Fix bwa problem."
```

I don't know about this - part of me likes it because the tutorials can be so clean and so complete, but part of me thinks it obsecures important things developers need to know to be effective.

I think regardless it is solid library functionality to add - potentially useful for things like bioconda recipes and the Bioconductor work.