Implement some ad-hoc documentation tests.
I just walked through the documentation and turned the examples into little bash scripts that can be executed and the result inspected to quickly test planemo-machine in an ongoing fashion. I intend to create a pre-GCC checklist for planemo-machine and I'll put reviewing these scripts against the docs on that.

I guess the next step would be actually annotate the docs - I should do that. Also these should probably be an Jupyter script instead of a bash script? Extract commands from RST into a Jupyter notebook - a bit that exists - or maybe we should go the other way. Who knows...