Implement external Galaxy engine.
This work is focused on "test" and "run" commands for Galaxy tools and workflows, though it should enable other combinations of artifacts and commands such as serving workflows and various CWL artifact operations against that fork of Galaxy.

The tool piece of this requires an unreleased version of galaxy-lib (https://github.com/galaxyproject/galaxy-lib/pull/91) and unmerged Galaxy modifications to expose a APIs for external tool testing (https://github.com/galaxyproject/galaxy/pull/5545). Workflows are oddly enough likely closer to working with this WIP.

Implements #592.
Implements #508.