Fix dependency resolvers and --biocontainers to work with cwltool.
By default, don't mess with cwltool's normal behavior to make it more like a Planemo configured Galaxy. However, if --no-container is specified go ahead and turn on Planemo's default dependency resolvers for the cwltool run. Also if --biocontainers is specified to Planemo - turn them on for cwltool as well (--beta-use-biocontainers).