Add --no_shed_install option.
@bgruening's original advice was to base this on whether --galaxy_admin_key was an actual admin key or not and just disable this functionality if not. My issue with this is that Planemo assumes a default admin key (that works with 'planemo serve' defaults), so if the admin key is wrong and required I think it should be a hard faliure not a silent behavior change. To put this another way, Planemo is tool development SDK, the correct assumption is that the target Galaxy is something it can just install tools into without worry - if that is not a correct assumption you need to change the behavior of Planemo and to do that the --no_shed_install flag makes a lot of sense I think.

I understand the critique that there are too many commands with too many options, and I have done things to make the default behaviors better (for instance none of the --conda_*** flags really need to be used for typical runs now). Rather than restricting functionality or making too clever assumptions I prefer the to go the route of documentation that focused on the actual commands to run and the relevant options for that use case. Therefore I've updated the documentation related to external Galaxy testing - and that doesn't cover twenty options - it covers only three - the most important three for external Galaxy testing including this new one.