Make symlinks in tool tree work for planemo test
Fixes https://github.com/galaxyproject/planemo/issues/857 and https://github.com/galaxyproject/planemo/issues/865
Symlinks were problematic when the symlink pointed out of the tool
directory. This was not a problem when uploading to the tool shed,
because we'd dereference the symlinks when creating the repository
archive. So here we just copy the tool tree with shutil.copytree,
which dereferences symlinks by default.