support Singularity versions that lack the --env flag.
Prior to version 3.6, Singularity did not support the --env flag and instead relied on environment variables prefixed with SINGULARITYENV_ to set container environments. These environment variables are essential during the initialization of the PostgreSQL database. This update ensures compatibility with both newer versions of Singularity that support the --env flag and older versions that do not.