start implementing support for downloading output_data after an invocation completes
Currently, the --download_outputs flag can be used with planemo run, but it requires the user to wait until the execution is complete. If the user prefers to use --nowait or if the connection is lost during execution, they lose the ability to retrieve the outputs. This pull request begins to introduce support for downloading outputs after the invocation has finished, regardless of when or how it completes.